+++ Modules/Microsoft365DSC/Modules/M365DSCReverse.psm1
@@ -92,11 +92,18 @@ function Start-M365DSCConfigurationExtract
 
         [Parameter()]
         [Switch]
-        $Validate
+        $Validate,
+
+        [Parameter()]
+        [Switch]
+        $Parallel,
+
+        [Parameter()]
+        [System.Collections.Generic.Dictionary[System.String, System.Object]]
+        $ResourceSettings
     )
 
-    # Start by checking to see if a new Version of the tool is available in the
-    # PowerShell Gallery
+    # Start by checking to see if a new version of the tool is available in the PowerShell Gallery
     try
     {
         Write-Verbose -Message 'Testing Module Validity'
@@ -106,6 +113,7 @@ function Start-M365DSCConfigurationExtract
     {
         Add-M365DSCEvent -Message $_ -Source 'M365DSCReverse::Test-M365DSCModuleValidity'
     }
+
     try
     {
         $Global:PartialExportFileName = "$(New-Guid).partial.ps1"
@@ -509,16 +517,15 @@ function Start-M365DSCConfigurationExtract
             -Description 'Default Value Used to Ensure a Configuration Data File is Generated'
 
         Write-Verbose -Message 'Retrieving resources path'
-        $ResourcesPath = Join-Path -Path $PSScriptRoot `
+        $resourcesPath = Join-Path -Path $PSScriptRoot `
             -ChildPath '../DSCResources/' `
             -Resolve
         Write-Verbose -Message 'Loop through all resources files.'
-        $AllResources = Get-ChildItem $ResourcesPath -Recurse | Where-Object { $_.Name -like 'MSFT_*.psm1' }
+        $allResoures = Get-ChildItem $resourcesPath -Recurse | Where-Object { $_.Name -like 'MSFT_*.psm1' }
 
-        $i = 1
         $ResourcesToExport = @()
-        $ResourcesPath = @()
-        foreach ($ResourceModule in $AllResources)
+        $resourcesPath = @()
+        foreach ($ResourceModule in $allResoures)
         {
             try
             {
@@ -536,7 +543,7 @@ function Start-M365DSCConfigurationExtract
                         AuthenticationMethod = $authMethod.AuthMethod
                     }
                     $ResourcesToExport += $resourceInfo
-                    $ResourcesPath += $ResourceModule
+                    $resourcesPath += $ResourceModule
                 }
             }
             catch
@@ -591,71 +598,90 @@ function Start-M365DSCConfigurationExtract
             }
         }
 
-        $ResourcesPath = $ResourcesPath | Sort-Object $_.Name
-        foreach ($resource in $ResourcesPath)
-        {
+        Confirm-M365DSCDependencies
+        $partialExportName = $Global:PartialExportFileName
+        $resourcesPath = $resourcesPath | Sort-Object $_.Name
+        $synchronizedHashtable = [System.Collections.Hashtable]::Synchronized(@{
+            ResourceCounter = 1
+            ResourcesResult = @{}
+        })
+        $resourceDictionary = Get-M365DSCAllResourcesDictionary
+        $exportScriptBlock = {
+            $Global:PartialExportFileName = $using:partialExportName
+            $Global:M365DSCSkipDependenciesValidation = $true
+            $resource = $_
+            Set-M365DSCAllResourcesDictionary -DscResourceDictionary $using:resourceDictionary
             $resourceName = $resource.Name.Split('.')[0] -replace 'MSFT_', ''
-            $mostSecureAuthMethod = ($allSupportedResourcesWithMostSecureAuthMethod | Where-Object { $_.Resource -eq $resourceName }).AuthMethod
+            $mostSecureAuthMethod = ($using:allSupportedResourcesWithMostSecureAuthMethod | Where-Object { $_.Resource -eq $resourceName }).AuthMethod
 
             Import-Module $resource.FullName | Out-Null
-            $FilterExists = (Get-Command 'Export-TargetResource').Parameters.Keys.Contains('Filter')
+            $filterExists = (Get-Command 'Export-TargetResource').Parameters.Keys.Contains('Filter')
 
             $parameters = @{}
             switch ($mostSecureAuthMethod)
             {
                 { $_ -in 'CertificateThumbprint', 'CertificatePath', 'ApplicationSecret' }
                 {
-                    $parameters.Add('ApplicationId', $ApplicationId)
-                    $parameters.Add('TenantId', $TenantId)
+                    $parameters.Add('ApplicationId', $using:ApplicationId)
+                    $parameters.Add('TenantId', $using:TenantId)
                 }
                 'CertificateThumbprint'
                 {
-                    $parameters.Add('CertificateThumbprint', $CertificateThumbprint)
+                    $parameters.Add('CertificateThumbprint', $using:CertificateThumbprint)
                 }
                 'CertificatePath'
                 {
-                    $parameters.Add('CertificatePath', $CertificatePath)
-                    $parameters.Add('CertificatePassword', $CertificatePassword)
+                    $parameters.Add('CertificatePath', $using:CertificatePath)
+                    $parameters.Add('CertificatePassword', $using:CertificatePassword)
                 }
                 'ApplicationSecret'
                 {
-                    $applicationSecretValue = New-Object System.Management.Automation.PSCredential ('ApplicationSecret', (ConvertTo-SecureString $ApplicationSecret -AsPlainText -Force))
+                    $applicationSecretValue = New-Object System.Management.Automation.PSCredential ('ApplicationSecret', (ConvertTo-SecureString $using:ApplicationSecret -AsPlainText -Force))
                     $parameters.Add('ApplicationSecret', $applicationSecretValue)
                 }
                 { $_ -in 'Credentials', 'CredentialsWithApplicationId' }
                 {
-                    if ($AuthMethods -contains 'CredentialsWithApplicationId')
+                    if ($using:AuthMethods -contains 'CredentialsWithApplicationId')
                     {
-                        $parameters.Add('ApplicationId', $ApplicationId)
+                        $parameters.Add('ApplicationId', $using:ApplicationId)
                     }
-                    $parameters.Add('Credential', $Credential)
+                    $parameters.Add('Credential', $using:Credential)
                 }
                 'CredentialsWithTenantId'
                 {
-                    $parameters.Add('Credential', $Credential)
-                    $parameters.Add('TenantId', $TenantId)
+                    $parameters.Add('Credential', $using:Credential)
+                    $parameters.Add('TenantId', $using:TenantId)
                 }
                 'ManagedIdentity'
                 {
-                    $parameters.Add('ManagedIdentity', $ManagedIdentity)
-                    $parameters.Add('TenantId', $TenantId)
+                    $parameters.Add('ManagedIdentity', $using:ManagedIdentity)
+                    $parameters.Add('TenantId', $using:TenantId)
                 }
                 'AccessTokens'
                 {
-                    $parameters.Add('AccessTokens', $AccessTokens)
-                    $parameters.Add('TenantId', $TenantId)
+                    $parameters.Add('AccessTokens', $using:AccessTokens)
+                    $parameters.Add('TenantId', $using:TenantId)
                 }
             }
 
-            if ($ComponentsToSkip -notcontains $resourceName)
+            if ($using:ComponentsToSkip -notcontains $resourceName)
             {
-                Write-M365DSCHost -Message "[$i/$($ResourcesToExport.Length)] Extracting [" -DeferWrite
+                if ($PSVersionTable.PSEdition -eq 'Core')
+                {
+                    $module = Get-Module PSDesiredStateConfiguration
+                    if ($null -eq $module)
+                    {
+                        Import-Module -Name "PSDesiredStateConfiguration" -Global -Prefix 'Pwsh' -RequiredVersion 2.0.7
+                    }
+                }
+                $counter = ($using:synchronizedHashtable).ResourceCounter++
+                Write-M365DSCHost -Message "[$counter/$($using:ResourcesToExport.Length)] Extracting [" -DeferWrite
                 Write-M365DSCHost -Message $resourceName -ForegroundColor Green -DeferWrite
                 Write-M365DSCHost -Message '] using {' -DeferWrite
                 Write-M365DSCHost -Message $mostSecureAuthMethod -ForegroundColor Cyan -DeferWrite
                 Write-M365DSCHost -Message '}...' -DeferWrite
-                $exportString = [System.Text.StringBuilder]::New()
-                if ($GenerateInfo)
+                $exportString = [System.Text.StringBuilder]::new()
+                if ($using:GenerateInfo)
                 {
                     $exportString.Append("`r`n        # For information on how to use this resource, please refer to:`r`n") | Out-Null
                     $exportString.Append("        # https://github.com/microsoft/Microsoft365DSC/wiki/$($resource.Name.Split('.')[0] -replace 'MSFT_', '')`r`n") | Out-Null
@@ -664,10 +690,10 @@ function Start-M365DSCConfigurationExtract
                 # Check if filters for the current resource were specified.
                 $resourceFilter = $null
                 $resourceName = $resource.Name.Split('.')[0] -replace 'MSFT_', ''
-                if ($FilterExists -and $null -ne $Filters -and $Filters.Keys.Contains($resourceName))
+                if ($filterExists -and $null -ne $using:Filters -and ($using:Filters).Keys.Contains($resourceName))
                 {
-                    $resourceFilter = $Filters.$resourceName
-                    if ($FilterExists)
+                    $resourceFilter = ($using:Filters).$resourceName
+                    if ($filterExists)
                     {
                         $parameters.Add('Filter', $resourceFilter)
                     }
@@ -678,9 +704,36 @@ function Start-M365DSCConfigurationExtract
                 }
                 $Global:M365DSCExportResourceTypes += $resourceName
                 $exportString.Append((Export-TargetResource @parameters)) | Out-Null
-                $i++
             }
-            $DSCContent.Append($exportString.ToString()) | Out-Null
+            ($using:synchronizedHashtable).ResourcesResult.Add($resourceName, $exportString.ToString())
+        }
+
+        if ($Parallel)
+        {
+            Write-M365DSCHost -Message "Starting export in parallel mode. Initialiation may take a while..."
+            $requiredModules = [System.Collections.Generic.List[System.String]]::new(50)
+            foreach ($resource in $ResourcesToExport)
+            {
+                foreach ($module in $resourceSettings[$resource.Name])
+                {
+                    if (-not $requiredModules.Contains($module))
+                    {
+                        $requiredModules.Add($module)
+                    }
+                }
+            }
+            $resourcesPath | Invoke-Parallel -ScriptBlock $exportScriptBlock -ModuleName $requiredModules -Verbose
+        }
+        else
+        {
+            Write-M365DSCHost -Message "Starting export in sequential mode..."
+            $exportScriptBlock = [ScriptBlock]::Create($exportScriptBlock.ToString().Replace('$using:', '$'))
+            $resourcesPath | ForEach-Object -Process $exportScriptBlock
+        }
+
+        foreach ($resource in $($synchronizedHashtable.Keys | Sort-Object))
+        {
+            $DSCContent.Append($synchronizedHashtable.ResourcesResult.$resource) | Out-Null
         }
 
         # Close the Node and Configuration declarations